portfolio_backend/requirements.txt
djangorestframework
django-cors-headers
Pillow
